MD server support using DHCP agent and MD Proxy
In Quantum's metadata solution for Grizzly in 'isolate mode', that is
without an l3 agent, the dhcp agent spawns a metadata proxy for each
network. The metadata proxy forwards them to the metadata agent using
a socket and adding in the http header the network id or the router id.
Midonet plugin uses DhcpDriverNoOp since the DHCP will be handled by the
controller. MidonetInterfaceDriver plugs the metadata proxy into
MidoNet. Every VM will receive a route through DHCP option 121
indicating the metadata proxy address as next hop for the MD server.
Every router will have a route for each subnet  to send the traffic
for the MD server to the metadata proxy.

Change-Id: I9cd55255bc338fa136d9d6c6b3226ad1801358b0
Signed-off-by: Rossella Sblendido <rossella@midokura.jp>