Change create_port to use ports' tenant ID
This patch changes MidoNetPluginV2.create_port to use the tenant ID
retrieved from the port given as an argument instead of the tenant ID of
the bridge associated with the port.

Neutron can have the "shared network", which belongs to the different
tenant from the current users' one. In this case the shared network is
translated into a bridge in MidoNet but it can belongs to another
tenant, "service" for instance. This confuses MidoNet and causes errors
when we attach or detach the security groups because it can't get the
chain across the tenant in MidoNet model.

I tested this patch attaching/detaching the security groups to/from the
instance bound to the regular network and the shared network. I
confirmed the security group was working appropriately with ping.

This fixes MN-1839.

Signed-off-by: Taku Fukushima <tfukushima@midokura.com>